B2XKlaim-BackEnd/src/test/java/com/example/B2XKlaim/ControllerTest/TranslationControllerTest.java
package com.example.B2XKlaim.ControllerTest;

import com.example.B2XKlaim.Controller.TranslationController;
import com.example.B2XKlaim.Service.Parser.BpmnParser;
import com.example.B2XKlaim.Service.bpmnElements.BpmnElements;
import com.example.B2XKlaim.Service.codeGenerator.Generator;
import com.example.B2XKlaim.Service.codeGenerator.Optimizer;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.http.ResponseEntity;
import java.util.*;
import org.springframework.http.HttpStatus;

public class TranslationControllerTest {

    @Mock
    private BpmnParser mockBpmnParser;

    @Mock
    private Generator mockGenerator;

    @Mock
    private Optimizer mockOptimizer;

    @InjectMocks
    private TranslationController translationController;

    @BeforeEach
    public void setup() {
        // Initialize mocks
        mockBpmnParser = mock(BpmnParser.class);
        mockGenerator = mock(Generator.class);
        mockOptimizer = mock(Optimizer.class);
        translationController = new TranslationController();
    }

    @Test
    public void testGenerateCode() throws Exception {
        // Setup test data
        String testBpmnXml = "...";
        BpmnElements mockBpmnElements = new BpmnElements(); // You'd populate this as necessary
        List<String> mockCollabTranslation = Arrays.asList("line1", "line2");
        List<String> mockOptimizedCollabo = Arrays.asList("optLine1", "optLine2");

        // Mocking behavior
        when(mockBpmnParser.parse(testBpmnXml)).thenReturn(mockBpmnElements);
        when(mockGenerator.translateBpmnCollaboration()).thenReturn(mockCollabTranslation);
        when(mockOptimizer.optimize(mockCollabTranslation)).thenReturn(mockOptimizedCollabo);
        // ...

        // Call the method
        ResponseEntity<Map<String, Object>> response = translationController.generateCode(testBpmnXml);

        // Asserts
        assertEquals(HttpStatus.OK, response.getStatusCode());
        Map<String, Object> responseBody = response.getBody();
        assertNotNull(responseBody);
        assertEquals("optLine1\noptLine2", responseBody.get("collaboration"));
        // ...
    }

    // For more test cases
}